pyproject.toml
[project]
name = "bamboost-tui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "florez", email = "florez@ethz.ch" }
]
requires-python = ">=3.9"
dependencies = [
    "bamboost",
    "sqlalchemy>=2.0.37",
    "textual-dev>=1.7.0",
]

[project.scripts]
bamboost-tui = "bamboost_tui:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyright]
pythonVersion = "3.9"
pythonPlatform = "Linux"
typeCheckingMode = "basic"
reportOptionalMemberAccess = false
reportOptionalSubscript = false

[tool.bamboost.index]
isolated = true

[tool.bamboost.options]
mpi = true
sortTableKey = "created_at"

[tool.uv.sources]
bamboost = { path = "../bamboost@next", editable = true }
